requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
google-generativeai
python-dotenv
pydantic
python-jose[cryptography]
bcrypt